🔥 Remove template lib
