+++ .tuist-version
@@ -1,0 +1,1 @@
+3.37.0
