+++ App/Resources/Assets.xcassets/Contents.json
