+++ App/Resources/LaunchScreen.storyboard
