--- CHANGELOG.md
@@ -1,16 +1,0 @@
-# Changelog
-
-## Next
-
-- Remove code moved to [ACKategories](https://github.com/AckeeCZ/ACKategories) ([#75](https://github.com/AckeeCZ/iOS-MVVM-ProjectTemplate/pull/75), kudos to @olejnjak)
-
-## 1.1.0
-- Updates ([#74](https://github.com/AckeeCZ/iOS-MVVM-ProjectTemplate/pull/74), kudos to @olejnjak)
-    - new build settings defaults
-    - add app icon
-    - update Tuist plugin
-    - require Firebase 10.19+ as it introduced some breaking changes
-
-## 1.0.0
-
-- Prepare whole new template ([#72](https://github.com/AckeeCZ/iOS-MVVM-ProjectTemplate/pull/72), kudos to @olejnjak)
