+++ Modules/Assets/Resources/Assets.xcassets/Contents.json
