--- ProjectTemplate/.gitignore
@@ -1,65 +1,0 @@
-*/Settings.bundle/Licenses*
-
-# Created by http://www.gitignore.io
-
-### OSX ###
-.DS_Store
-.AppleDouble
-.LSOverride
-
-# Thumbnails
-._*
-
-# Files that might appear on external disk
-.Spotlight-V100
-.Trashes
-
-# Directories potentially created on remote AFP share
-.AppleDB
-.AppleDesktop
-Network Trash Folder
-Temporary Items
-.apdisk
-
-
-### Objective-C ###
-# Xcode
-#
-build/
-*.pbxuser
-!default.pbxuser
-*.mode1v3
-!default.mode1v3
-*.mode2v3
-!default.mode2v3
-*.perspectivev3
-!default.perspectivev3
-xcuserdata
-*.xccheckout
-*.moved-aside
-DerivedData
-*.hmap
-*.ipa
-*.xcuserstate
-*.xcodeproj
-*.xcworkspace
-
-# CocoaPods
-Pods/
-
-# Carthage
-Carthage/Build
-Carthage/Checkouts
-
-#Fastlane
-fastlane/report.xml
-fastlane/Preview.html
-fastlane/screenshots
-fastlane/test_output
-screenshots
-
-#build output
-/outputs
-
-# tuist
-Derived
