--- ProjectTemplate/.swiftlint.yml
@@ -1,42 +1,0 @@
-whitelist_rules:
-  - class_delegate_protocol
-  - closing_brace
-  - colon
-  - comma
-  - custom_rules
-  - cyclomatic_complexity
-  - empty_enum_arguments
-  - file_length
-  - force_cast
-  - force_unwrapping
-  - function_body_length
-  - trailing_semicolon
-  - trailing_whitespace
-  - vertical_whitespace
-  - weak_delegate
-  - overridden_super_call
-  - superfluous_disable_command
-included: # paths to include during linting. `--path` is ignored if present.
-  - ProjectTemplate
-  - Tests
-excluded: # paths to ignore during linting. Takes precedence over `included`.
-  - Source/Generated
-  - Carthage
-  - Pods
-force_cast: warning # implicitly
-function_body_length:
-  - 100 # warning
-  - 300 # error
-type_body_length:
-  - 300 # warning
-  - 500 # error
-file_length:
-  warning: 500
-  error: 1200
-
-custom_rules:
-  todo_redmine:
-    name: "TODO Redmine"
-    regex: 'TODO: (?!https:\/\/redmine\.ack\.ee\/issues\/\d)'
-    message: "TODO must contain the full link to the Redmine ticket"
-    severity: warning
