--- ProjectTemplate/.tuist-version
@@ -1,1 +1,0 @@
-3.36.5
