+++ Tuist/ProjectDescriptionHelpers/Extensions/InfoPlistExtensions.swift
