+++ Tuist/ProjectDescriptionHelpers/Targets/Assets.swift
