+++ Tuist/ProjectDescriptionHelpers/Targets/UserManager.swift
